web3 rpc interface implements

+++ components/contracts/modules/ethereum/tests/ethereum_integration.rs
@@ -159,7 +159,7 @@ fn test_mint_balance(who: &Address, balance: u128, height: u64) {
     let ctx = BASE_APP
         .lock()
         .unwrap()
-        .create_query_context(height as i64, false)
+        .create_query_context(height, false)
         .unwrap();
     assert_eq!(module_account::App::<BaseApp>::balance(&ctx, who), balance);
 }
